1D/2D Spectra of pair of SiO2 and Al2O3
This was an initial investigation to try and see which DBR materials would be best to investigate for our experiment. The goal of our experiment is to build off of the results from the T. Schwarz paper and instead of using a simple Ag mirror we hope to increase the Q factor by several fold using DBRs.